Refactor data layer
