+++ app/src/main/java/ru/neexol/rtut/data/remote/lessons/LessonsRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package ru.neexol.rtut.data.remote
+package ru.neexol.rtut.data.remote.lessons
 
 import javax.inject.Inject
