+++ app/src/main/java/ru/neexol/rtut/data/remote/lessons/models/GroupLessonsResponse.kt
@@ -1,4 +1,4 @@
-package ru.neexol.rtut.data.remote.models
+package ru.neexol.rtut.data.remote.lessons.models
 
 import ru.neexol.rtut.domain.models.GroupLessons
